+++ CHANGELOG.md
@@ -7,9 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## - Unreleased
 
-- Update IconClose
+- Update IconClose.
 
-- Update styles component Modal
+- Update Modal styles.
 
 ## [0.21.0] - 2020-08-11